adds transak connection panel and refactor panes into components
