+++ packages/desktop/views/dashboard/buy-sell/components/index.ts
@@ -1,0 +1,3 @@
+export { default as TransakAccountPanel } from './TransakAccountPanel.svelte'
+export { default as TransakConnectionPanel } from './TransakConnectionPanel.svelte'
+export { default as TransakInfoPanel } from './TransakInfoPanel.svelte'
